Latest Update

+++ GreenBasket.Application/DTOs/Products/ProductDto.cs
@@ -9,7 +9,7 @@ namespace GreenBasket.Application.DTOs.Products
         public decimal Price { get; set; }
         public string Unit { get; set; } = string.Empty;
         public bool Organic { get; set; }
-        public int StockQty { get; set; }
+        public decimal StockQty { get; set; }
         public string StockStatus { get; set; } = string.Empty;
 
         public string? FarmOrigin { get; set; }
